IDENTITY-5484: Code Refactoring
