--- components/org.wso2.carbon.identity.sp.metadata.saml2/src/main/java/org/wso2/carbon/identity/sp/metadata/saml2/Exception/InvalidMetadataException.java
@@ -1,17 +1,0 @@
-package org.wso2.carbon.identity.sp.metadata.saml2.Exception;
-
-/**
- * Created by pasindu on 11/10/16.
- */
-public class InvalidMetadataException extends Exception {
-
-    public InvalidMetadataException(String message) {
-        super(message);
-    }
-
-    public InvalidMetadataException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-
-}
